bring back `vl.fieldDef.field()`
