+++ src/compiler/Model.ts
@@ -6,6 +6,7 @@ import {instantiate} from '../schema/schemautil';
 import {COLUMN, ROW, X, Y, COLOR, DETAIL, Channel, supportMark} from '../channel';
 import {SOURCE, SUMMARY} from '../data';
 import * as vlFieldDef from '../fielddef';
+import {FieldRefOption} from '../fielddef';
 import * as vlEncoding from '../encoding';
 import {compileLayout} from './layout';
 import {AREA, BAR, POINT, TICK, CIRCLE, SQUARE, Mark} from '../mark';
@@ -14,24 +15,8 @@ import * as schemaUtil from '../schema/schemautil';
 import {StackProperties} from './stack';
 import {type as scaleType} from './scale';
 import {getFullName, NOMINAL, ORDINAL, TEMPORAL} from '../type';
-import {contains, duplicate, isArray} from '../util';
+import {contains, duplicate, extend, isArray} from '../util';
 import {Encoding} from '../schema/encoding.schema';
-
-
-interface FieldRefOption {
-  /** exclude bin, aggregate, timeUnit */
-  nofn?: boolean;
-  /** exclude aggregation function */
-  noAggregate?: boolean;
-  /** include 'datum.' */
-  datum?: boolean;
-  /** replace fn with custom function prefix */
-  fn?: string;
-  /** prepend fn with custom function prefix */
-  prefn?: string;
-  /** append suffix to the field ref for bin (default='_start') */
-  binSuffix?: string;
-}
 
 
 /**
@@ -169,30 +154,15 @@ export class Model {
     return this._spec.encoding[channel];
   }
 
-  // get "field" reference for vega
-  public field(channel: Channel, opt?: FieldRefOption) {
-    opt = opt || {};
-
+  /** Get "field" reference for vega */
+  public field(channel: Channel, opt: FieldRefOption = {}) {
     const fieldDef = this.fieldDef(channel);
-
-    var f = (opt.datum ? 'datum.' : '') + (opt.prefn || ''),
-      field = fieldDef.field;
-
-    if (vlFieldDef.isCount(fieldDef)) {
-      return f + 'count';
-    } else if (opt.fn) {
-      return f + opt.fn + '_' + field;
-    } else if (!opt.nofn && fieldDef.bin) {
-      var binSuffix = opt.binSuffix ||
-        (scaleType(fieldDef, channel) === 'ordinal' ? '_range' : '_start');
-      return f + 'bin_' + field + binSuffix;
-    } else if (!opt.nofn && !opt.noAggregate && fieldDef.aggregate) {
-      return f + fieldDef.aggregate + '_' + field;
-    } else if (!opt.nofn && fieldDef.timeUnit) {
-      return f + fieldDef.timeUnit + '_' + field;
-    } else {
-      return f + field;
+    if (fieldDef.bin) { // bin has default suffix that depends on scaleType
+      opt = extend({
+        binSuffix: scaleType(fieldDef, channel) === 'ordinal' ? '_range' : '_start'
+      }, opt);
     }
+    return vlFieldDef.field(fieldDef, opt);
   }
 
   public fieldTitle(channel: Channel): string {
