+++ src/fielddef.ts
@@ -5,6 +5,40 @@ import {contains, getbins} from './util';
 import {NOMINAL, ORDINAL, QUANTITATIVE, TEMPORAL} from './type';
 
 
+export interface FieldRefOption {
+  /** exclude bin, aggregate, timeUnit */
+  nofn?: boolean;
+  /** exclude aggregation function */
+  noAggregate?: boolean;
+  /** include 'datum.' */
+  datum?: boolean;
+  /** replace fn with custom function prefix */
+  fn?: string;
+  /** prepend fn with custom function prefix */
+  prefn?: string;
+  /** append suffix to the field ref for bin (default='_start') */
+  binSuffix?: string;
+}
+
+export function field(fieldDef: FieldDef, opt: FieldRefOption = {}) {
+  var f = (opt.datum ? 'datum.' : '') + (opt.prefn || ''),
+    field = fieldDef.field;
+
+  if (isCount(fieldDef)) {
+    return f + 'count';
+  } else if (opt.fn) {
+    return f + opt.fn + '_' + field;
+  } else if (!opt.nofn && fieldDef.bin) {
+    return f + 'bin_' + field + opt.binSuffix;
+  } else if (!opt.nofn && !opt.noAggregate && fieldDef.aggregate) {
+    return f + fieldDef.aggregate + '_' + field;
+  } else if (!opt.nofn && fieldDef.timeUnit) {
+    return f + fieldDef.timeUnit + '_' + field;
+  } else {
+    return f + field;
+  }
+}
+
 // TODO remove these "isDimension/isMeasure" stuff
 function _isFieldDimension(fieldDef: FieldDef) {
   return contains([NOMINAL, ORDINAL], fieldDef.type) || !!fieldDef.bin ||
